requirements.txt
fastapi 
uvicorn 
langchain 
langgraph 
openai 
faiss-cpu 
chromadb 
pydantic
langchain-community
langchain-together
python-dotenv
python-multipart